Admin Refinement

+++ src/utils/backend/person/student.ts
@@ -14,7 +14,7 @@ import { supabase } from "@utils/supabaseClient";
 import { ClassWNumber } from "@utils/types/class";
 import { BackendReturn } from "@utils/types/common";
 import { ImportedStudentData, Student } from "@utils/types/person";
-import { ClassroomDB } from "@utils/types/database/class";
+import { ClassroomDB, ClassroomTable } from "@utils/types/database/class";
 import {
   PersonDB,
   PersonTable,
@@ -22,7 +22,7 @@ import {
 } from "@utils/types/database/person";
 
 const prefixMap = {
-  เด็กชาย: "Master.",
+  เด็กชาย: "Master",
   นาย: "Mr.",
   นาง: "Mrs.",
   นางสาว: "Miss.",
@@ -32,7 +32,7 @@ export async function createStudent(
   student: Student,
   email: string,
   isAdmin: boolean = false
-): Promise<{ data: StudentTable[] | null; error: PostgrestError | null }> {
+): Promise<{ data: StudentTable | null; error: PostgrestError | null }> {
   const { data: person, error: personCreationError } = await createPerson(
     student
   );
@@ -46,7 +46,8 @@ export async function createStudent(
     .insert({
       person: person[0]?.id,
       std_id: student.studentID.trim(),
-    });
+    })
+    .single();
   if (studentCreationError || !createdStudent) {
     console.error(studentCreationError);
     // delete the created person
@@ -57,6 +58,35 @@ export async function createStudent(
     return { data: null, error: studentCreationError };
   }
 
+  // add the student to the students array of the class
+  // get the class
+  const { data: classData, error: classError } = await supabase
+    .from<ClassroomTable>("classroom")
+    .select("*")
+    .match({ number: student.class.number })
+    .limit(1)
+    .single();
+  if (classData) {
+    // add the student to the students array of the class
+    // console.log(classData);
+    const { data: updatedClass, error: updateClassError } = await supabase
+      .from<ClassroomTable>("classroom")
+      .update({
+        students: [...classData.students, createdStudent.id],
+        // put student at the index of classNo - 1
+        no_list: [
+          ...classData.no_list.slice(0, student.classNo - 1),
+          createdStudent.id,
+          ...classData.no_list.slice(student.classNo),
+        ],
+      })
+      .match({ number: student.class.number, year: getCurrentAcedemicYear() });
+
+    if (updateClassError || !updatedClass) {
+      console.error(updateClassError);
+    }
+  }
+
   // register an account for the student
   const res = await fetch("/api/account/student", {
     method: "POST",
@@ -65,8 +95,7 @@ export async function createStudent(
     },
     body: JSON.stringify({
       email,
-      password: student.birthdate.split("-").join(""),
-      id: createdStudent[0]?.id,
+      id: createdStudent.id,
       isAdmin,
     }),
   });
@@ -84,18 +113,17 @@ export async function deleteStudent(student: Student) {
 
   if (selectingError) {
     console.error(selectingError);
-    return;
   }
 
   if (!userid) {
     console.error("No user found");
-    return;
   }
 
   const { data: deleting, error } = await supabase
     .from<StudentTable>("student")
     .delete()
-    .match({ id: student.id });
+    .match({ id: student.id })
+    .single();
   if (error || !deleting) {
     console.error(error);
     return;
@@ -105,13 +133,25 @@ export async function deleteStudent(student: Student) {
   const { data: person, error: personDeletingError } = await supabase
     .from<PersonTable>("people")
     .delete()
-    .match({ id: deleting[0].person });
+    .match({ id: deleting.person });
 
   if (personDeletingError || !person) {
     console.error(personDeletingError);
     return;
   }
 
+  // remove the student from the class
+  const { data: classStudent, error: classStudentError } = await supabase
+    .from("classroom")
+    .update({
+      students: supabase.rpc("array_remove", ["students", student.id]),
+    })
+    .in("students", [student.id]);
+
+  if (classStudentError || !classStudent) {
+    console.error(classStudentError);
+    return;
+  }
   // Delete account of the student
   await fetch(`/api/account`, {
     method: "DELETE",
@@ -119,7 +159,7 @@ export async function deleteStudent(student: Student) {
       "Content-Type": "application/json",
     },
     body: JSON.stringify({
-      id: userid.id,
+      id: userid?.id,
     }),
   });
 }
@@ -131,7 +171,7 @@ export async function importStudents(data: ImportedStudentData[]) {
         id: 0,
         prefix: {
           th: student.prefix,
-          "en-US": prefixMap[student.prefix]
+          "en-US": prefixMap[student.prefix],
         },
         role: "student",
         name: {
@@ -152,20 +192,32 @@ export async function importStudents(data: ImportedStudentData[]) {
         contacts: [],
         class: {
           id: 0,
-          number: student.class_number,
+          number: student.class,
         },
-        classNo: 0,
+        classNo: student.class_number,
       };
       const email = student.email;
       return { person, email };
     }
   );
+  // console.log(students);
 
-  await Promise.all(
-    students.map(
-      async (student) => await createStudent(student.person, student.email)
-    )
-  );
+  // await Promise.all(
+  //   students.map(
+  //     async (student) => await createStudent(student.person, student.email)
+  //   )
+  // );
+
+  // for (const student of students) {
+  //   await createStudent(student.person, student.email);
+  // }
+  // sequentially create students
+  for (let i = 0; i < students.length; i++) {
+    await createStudent(students[i].person, students[i].email);
+    if (i % 10 === 0) {
+      console.log(i);
+    }
+  }
 }
 
 export async function getClassOfStudent(
